+++ app/assets/rive/off_road_car.riv
